feat: moved trigger to webhook direction
