+++ webhook/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "storage-service",
+    "version": "1.0.0",
+    "main": "index.js",
+    "scripts": {
+      "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "description": "",
+    "dependencies": {
+      "@aws-sdk/client-s3": "^3.701.0",
+      "cors": "^2.8.5",
+      "dotenv": "^16.4.5",
+      "express": "^4.21.1",
+      "morgan": "^1.10.0",
+      "multer": "^1.4.5-lts.1",
+      "multer-s3": "^3.0.1",
+      "uuid": "^11.0.3"
+    }
+  }
+
